% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/private.R
\name{private}
\alias{private}
\alias{\%private\%}
\alias{\%private\%<-}
\title{Define and access private functions}
\usage{
parent \%private\% child

`\%private\%`(parent, child) <- value
}
\arguments{
\item{parent}{The \code{"private_parent"} function, or the target to be converted to a \code{"private_parent"}}

\item{child}{The name of the object to access or to define in \code{parent}'s private namespace}

\item{value}{An object, often a function.}
}
\description{
These functions provide a convenient way to create namespaces associated with
a given function. This avoids cluttering the main namespace of a package,
prevents name collision and makes the scope of helper functions explicit in the code.
}
\details{
The first time \verb{\%private\%<-} is called on a function, its environment (usually
the package's namespace) is changed to be a new namespace, child of its original environment.

Anytime \verb{\%private\%<-} is called on a function its environment is populated with
the object defined in the right hand side, which will be bound there and have its environment set there as well
if it's a function.

\verb{\%private\%} is called to access the child function from outside, to test it
for instance.
}
